[rollbar] fix broken rollbar setup
We noticed that we haven't been getting rollbars for a while. While we'd
like to think that's because we fixed every last bug, we know ourselves
better than that

After some investigation, it seems we’re configuring rollbar an OLD way, via this:

https://docs.rollbar.com/docs/laravel#:~:text=%24this%2D%3Eapp%2D%3Eregister(%5CRollbar%5CLaravel%5CRollbarServiceProvider%3A%3Aclass)%3B

which is indicated for “laravel 5.5 and lower”

Adding it as a logger (top of that page) works swimmingly.

I think this is from the rollbar/rollbar v2.0.0 -> v2.1.0 upgrade we did back in june: https://github.com/snipe/snipe-it/blame/aa3f94973f5efe595bdc2adb411666dcdf3b427f/composer.lock#L7894-L7895

THAT BEING SAID… there’s nothing about it in the changelog… https://github.com/rollbar/rollbar-php/releases/tag/v2.1.0.   but then there’s ~nothing in the changelog for 2.1.0 even though there are 230+ commits :eyeroll:

INTERESTINGLY, removing The Old Way also made it not work, so I'm
leaving both in for this PR